move alert machinery to a util file
