+++ choreolib/src/main/java/choreo/auto/AutoTrajectory.java
@@ -5,14 +5,15 @@ package choreo.auto;
 import static edu.wpi.first.wpilibj.Alert.AlertType.kError;
 
 import choreo.Choreo;
-import choreo.Choreo.MultiAlert;
 import choreo.Choreo.TrajectoryLogger;
 import choreo.auto.AutoFactory.AutoBindings;
 import choreo.trajectory.DifferentialSample;
 import choreo.trajectory.SwerveSample;
 import choreo.trajectory.Trajectory;
 import choreo.trajectory.TrajectorySample;
+import choreo.util.AlertUtil;
 import choreo.util.AllianceFlipUtil;
+import choreo.util.AlertUtil.MultiAlert;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
@@ -43,17 +44,17 @@ public class AutoTrajectory {
   // and far between. This helps with more novice users
 
   private static final MultiAlert triggerTimeNegative =
-      Choreo.multiAlert(causes -> "Trigger time cannot be negative for " + causes, kError);
+      AlertUtil.multiAlert(causes -> "Trigger time cannot be negative for " + causes, kError);
   private static final MultiAlert triggerTimeAboveMax =
-      Choreo.multiAlert(
+      AlertUtil.multiAlert(
           causes -> "Trigger time cannot be greater than total trajectory time for " + causes + ".",
           kError);
   private static final MultiAlert eventNotFound =
-      Choreo.multiAlert(causes -> "Event Markers " + causes + " not found.", kError);
+      AlertUtil.multiAlert(causes -> "Event Markers " + causes + " not found.", kError);
   private static final MultiAlert noSamples =
-      Choreo.multiAlert(causes -> "Trajectories " + causes + " have no samples.", kError);
+      AlertUtil.multiAlert(causes -> "Trajectories " + causes + " have no samples.", kError);
   private static final MultiAlert noInitialPose =
-      Choreo.multiAlert(
+      AlertUtil.multiAlert(
           causes -> "Unable to get initial pose for trajectories " + causes + ".", kError);
 
   private static final double DEFAULT_TOLERANCE_METERS = Units.inchesToMeters(3);
